20-Projet_IA_MBT/00-BI_auto_Python/requirements.txt
gradio
python-docx
docx2pdf
psycopg2-binary
python-dotenv